backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
asyncio==3.4.3
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
av==12.3.0
bidict==0.23.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
constantly==23.10.4
cryptography==44.0.0
DateTime==5.5
distlib==0.3.9
distro==1.9.0
Django==5.1.4
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.15.2
djangorestframework-jsonapi==6.0.0
dnspython==2.7.0
eventlet==0.38.2
filelock==3.16.1
frozenlist==1.5.0
gevent==24.11.1
gevent-websocket==0.10.1
google-crc32c==1.6.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
hyperlink==21.0.0
icalendar==6.0.1
icnsutil==1.1.0
idna==3.10
ifaddr==0.2.0
incremental==24.7.2
inflection==0.5.1
jiter==0.6.1
multidict==6.1.0
netifaces==0.10.6
numpy==2.1.2
openai==1.52.2
pbr==6.1.0
pillow==11.0.0
platformdirs==4.3.6
propcache==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.1.1
pylibsrtp==0.10.0
pyOpenSSL==24.3.0
pyperclip==1.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
python-engineio==4.11.1
python-socketio==5.11.4
pytz==2022.6
requests==2.32.3
service-identity==24.2.0
setuptools==75.6.0
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
socketio==0.2.1
socksio==1.0.0
sqlparse==0.5.3
stevedore==5.4.0
tqdm==4.66.5
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2022.6
uncurl==0.0.11
urllib3==2.2.3
uuid==1.30
virtualenv==20.28.0
virtualenv-clone==0.5.7
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
